change src2

+++ src/main/java/ru/netology/javaqa/javaqamvn/services/StatsService.java
@@ -1,7 +1,7 @@
 package ru.netology.javaqa.javaqamvn.services;
 
 public class StatsService {
-    public int minSales(int[] sales) {
+    public int minSales(long[] sales) {
         int minMonth = 0; // номер месяца с минимальными продажами среди просмотренных ранее
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] <= sales[minMonth]) { // значит, в рассматриваемом i-м месяце продаж меньше
@@ -11,7 +11,7 @@ public class StatsService {
         return minMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
 
-    public int maxSales(int[] sales) {
+    public int maxSales(long[] sales) {
         int maxMonth = 0; // номер месяца с минимальными продажами среди просмотренных ранее
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] >= sales[maxMonth]) { // значит, в рассматриваемом i-м месяце продаж меньше
@@ -20,52 +20,55 @@ public class StatsService {
         }
         return maxMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
-public int findAmountSales(int[] sales) {
-        int amountSales = 0;
+
+    public long findAmountSales(long[] sales) {
+        long amountSales = 0;
         for (int i = 0; i < sales.length; i++) {
-           amountSales+=sales[i];
+            amountSales += sales[i];
         }
         return amountSales;
-}
-    public int averageAmountSalesPerMonth(int[] sales) {
-        int amountSales = 0;
+    }
+
+    public long averageAmountSalesPerMonth(long[] sales) {
+        long amountSales = 0;
         for (int i = 0; i < sales.length; i++) {
-            amountSales+=sales[i];
+            amountSales += sales[i];
         }
         amountSales = amountSales / sales.length;
         return amountSales;
     }
-    public int lowerAverage(int[] sales) {
-        int amountSales = 0;
+
+    public int lowerAverage(long[] sales) {
+        long amountSales = 0;
         for (int i = 0; i < sales.length; i++) {
-            amountSales+=sales[i];
+            amountSales += sales[i];
         }
         amountSales = amountSales / sales.length;
         int lowerMonth = 0;
-        for (int i = 0; i < sales.length; i++){
-            if (sales[i] < amountSales){
+        for (int i = 0; i < sales.length; i++) {
+            if (sales[i] < amountSales) {
                 lowerMonth++;
             }
         }
         return lowerMonth;
 
     }
-    public int higherAverage(int[] sales) {
-        int amountSales = 0;
+
+    public int higherAverage(long[] sales) {
+        long amountSales = 0;
         for (int i = 0; i < sales.length; i++) {
-            amountSales+=sales[i];
+            amountSales += sales[i];
         }
         amountSales = amountSales / sales.length;
         int higherMonth = 0;
-        for (int i = 0; i < sales.length; i++){
-            if (sales[i] > amountSales){
+        for (int i = 0; i < sales.length; i++) {
+            if (sales[i] > amountSales) {
                 higherMonth++;
             }
         }
         return higherMonth;
 
     }
-
 
 
 }
